Stub for multiplatform support
