+++ gontlm_other.go
@@ -1,0 +1,12 @@
+// +build !windows
+
+package main
+
+import (
+	"os"
+)
+
+func getProxyServer() (proxyServer string) {
+	proxyServer = os.Args[1]
+	return
+}
